requirements.txt
snowflake-connector-python
pandas
streamlit
google-cloud-storage